pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag_micro"
version = "0.1.0"
description = "RAG-пайплайн для технической документации (BM25 + Dense + Гибрид)"
authors = [{name = "nosignalx2k", email = "kingofsweetsx2k@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.24.0",
    "pdfplumber>=0.11.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "whoosh>=2.7.4",
    "faiss-cpu>=1.8.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.66.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.1",
    "rapidfuzz>=3.9.0",
    "tabulate>=0.9.0",
    "typer>=0.12.3",
    "pydantic>=2.6.0",
    "textdistance>=4.6.2",
    "llama-cpp-python>=0.2.84",
    "openai>=1.35.7",
]

[tool.setuptools.packages.find]
include = ["rag_micro*"]


[project.scripts]
rag-cli = "rag_micro.rag_cli:app"
